\name{DEscore}
\docType{data}
\alias{DEscore}
\alias{DSscore}
\alias{DEscore.perm}
\alias{DSscore.perm}
\title{Pre-calculated DE/DS scores}
\description{
\code{DEscore} and \code{DSscore} are pre-calculated DE and DS scores, respectively; 
\code{DEscore.perm} and \code{DSscore.perm} are pre-calculated DE and DS scores on the permutation data sets, respectively; 
They are used in examples of the \code{SeqGSEA} package. 
Note that these scores are of no meaning but to demonstrate the usage of functions. 
}
\usage{
data("DEscore")
data("DEscore.perm")
data("DSscore")
data("DSscore.perm")
}
\references{
Xi Wang and Murray J. Cairns (2013). Gene Set Enrichment Analysis of RNA-Seq Data: Integrating Differential Expression and Splicing. BMC Bioinformatics, in press
}
\keyword{datasets}
