\name{GS_example}
\docType{data}
\alias{GS_example}
\title{SeqGeneSet object example}
\description{
An exemplified SeqGeneSet object to demonstrate functions in the \code{SeqGSEA} package. 
This object was generated with collection #6 (C6) gene sets of the Molecular Signatures 
Database (MSigDB) v3.1. 
}
\usage{
data("GS_example")
}
\references{
Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B. L., Gillette, M. A., Paulovich, A., Pomeroy, S. L., Golub, T. R., Lander, E. S., and Mesirov, J. P. (2005). Gene set enrichment analysis: a knowledge-based approach for interpreting genome-wide expression profiles. Proc Natl Acad Sci USA, 102(43): 15545-50. 
}
\keyword{datasets}