\name{RCS_example}
\docType{data}
\alias{RCS_example}
\title{ReadCountSet object example}
\description{
An exemplified ReadCountSet object to demonstrate functions in the \code{SeqGSEA} package. 
This object is comprised of 20 samples across 5,000 exons, a part of the prostate cancer 
RNA-Seq data set from Kannan et al (2011). 
Please note that the count data in this example object is incomplete. 
}
\usage{
data("RCS_example")
}
\references{
Kannan, K., Wang, L., Wang, J., Ittmann, M. M., Li, W., and Yen, L. (2001). Recurrent chimeric RNAs enriched in human prostate cancer identified by deep sequencing. Proc Natl Acad Sci USA, 108(22): 9172-7. 
}
\keyword{datasets}